Add support for manu/dock integration, fix dependency resolution, rename pixbuf library
Add the ige-mac-integration library to the framework, and also setup
"quit" menu item and dock handling so that this works out of the box.

Fix a bug with dependency resolution, listing the files in Libraries
didn't find any files before.

Rename libgdk_pixbuf to libgdk-pixbuf since the underscore interferes
with the linker on darwin when using sublibraries.